requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
langchain
langchain-openai
langchain-community
pandas
openpyxl
pyngrok
nest-asyncio
pinecone-client
